Major refactor / reorg
