+++ frontend/src/admin/components/showcase/modals/IntroductionInitializedModal.tsx
