+++ frontend/src/admin/pages/ShowcasePage.tsx
@@ -5,9 +5,9 @@ import { useLocation, useNavigate, useSearchParams } from 'react-router-dom'
 import { adminBaseRoute } from '../api/adminApi'
 import { AdminNavbar } from '../components/AdminNavbar'
 import { SecondaryNavbar } from '../components/showcase/SecondaryNavbar'
-import { IntroductionTab } from '../components/showcase/tabs/IntroductionTab'
-import { PersonaTab } from '../components/showcase/tabs/PersonaTab'
-import { ScenariosTab } from '../components/showcase/tabs/ScenariosTab'
+import { IntroductionTab } from '../components/showcase/introduction/IntroductionTab'
+import { PersonaTab } from '../components/showcase/persona/PersonaTab'
+import { ScenariosTab } from '../components/showcase/scenarios/ScenariosTab'
 import { useShowcase } from '../hooks/useShowcase'
 
 type TabId = 'persona' | 'introduction' | 'scenarios'
